+++ src/hello.cc
@@ -1,7 +1,6 @@
 #include <iostream>
-#include "metron/hello.hh"
 
 void hello_greeting()
 {
-    std::cout << hello::greeting() << "\n";
+    std::cout << "Hello World!" << "\n";
 }